requirements.txt
admin-totals==1.0.1
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
configparser==5.0.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.6
django-admin-material==0.1.0
django-admin-rangefilter==0.6.0
django-import-export==2.1.0
django-material-admin==1.7.2
et-xmlfile==1.0.1
idna==2.9
jdcal==1.4.1
joblib==0.15.1
MarkupPy==1.14
mysql-connector==2.2.9
mysqlclient==1.4.6
natsort==7.0.1
numpy==1.18.4
odfpy==1.4.1
openpyxl==3.0.3
pandas==1.0.3
pandas-flavor==0.2.0
psycopg2-binary==2.8.5
pyjanitor==0.20.6
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
scikit-learn==0.23.1
scipy==1.4.1
six==1.14.0
sqlparse==0.3.1
tablib==2.0.0
threadpoolctl==2.0.0
urllib3==1.25.9
xarray==0.15.1
xlrd==1.2.0
xlwt==1.3.0
